feat(opossum-lib): implement Opossum file reader (ZIP format)
- OpossumFileReader implementing InputReader trait
- ZIP file reading with input.json (required) and output.json (optional)
- Conversion from OpossumInputFileModel to Opossum
- Resource tree building from flat path structure
- 7 integration tests with opossum_input.json fixture

+++ opossum-lib/src/input_formats/opossum/entities.rs
@@ -63,7 +63,7 @@ pub struct SourceInfoModel {
 #[serde(rename_all = "camelCase")]
 pub struct OpossumPackageModel {
     pub source: SourceInfoModel,
-    pub attribution_confidence: Option<i32>,
+    pub attribution_confidence: Option<f64>,
     pub comment: Option<String>,
     pub package_name: Option<String>,
     pub package_version: Option<String>,
